feat: Added error handling for friend and meet requests. Also added scheduled mail reminder backend in meet requests.

+++ socialite/server/meetrequest/meet.js
@@ -2,17 +2,71 @@ const data = require("../util/userdata");
 const { UserDets, Meet } = require("../models/User")
 
 const moment = require("moment-timezone")
-const schedule = require('node-schedule-tz')
 const { UserInputError } = require('apollo-server')
 
 async function meetaccept(user_id, fren_id)
 {
 	try{
+		errors = {}
 		var meet = await data.getMeet(user_id, fren_id);	
-		await data.updateAccRejMeet(user_id, fren_id, meet._id);	
+		dets = await data.getUserDetsById(user_id)
+		if(dets.sendmeet.includes(meet._id))
+		{
+			errors.general = 'You have already sent the request. Kindly refresh the page.';
+			throw new UserInputError('You have already sent the request. Kindly refresh the page.', { errors });
+		}
+
+		fdate = moment(meet.date).format("DD-MM-YYYY")
+		ftime = moment(moment(meet.time, "HH:mm:ss")).format("HH:mm:ss")
+		var fts = moment(`${fdate} ${ftime}`, 'DD-MM-YYYY HH:mm:ss').format();
+		fts = moment(fts)
+		var now = moment().format('YYYY-MM-DD HH:mm:ss')
+		now = moment(now)
+
+		if(now > fts)
+		{
+			errors.general = 'Invalid timestamp';
+			throw new UserInputError('Invalid timestamp', { errors });
+		}
+
+		await data.updateAccRejMeet(user_id, fren_id, meet._id)	
 		await data.newNotif(fren_id, user_id, "macc");
 		await data.removeNotif(user_id, fren_id, "mreq");
 		await data.newNotif(user_id, fren_id, "macc");
+
+		if(meet.notif)
+		{
+			title = "Meet Reminder"
+			var user1 = await data.getUserInfo(user_id)
+			var user2 = await data.getUserInfo(fren_id)
+			msg1 = `You have a meet scheduled with ${user2.email}(username: ${user2.username}) right now!
+				<br>Meet Details:
+				<br>Type:${meet.type}
+				<br>Duration:${meet.duration}
+				<br>Message:${meet.msg}
+				<br>Link:${meet.link}
+				<br>Place:${meet.place}
+			`
+			msg2 = `You have a meet scheduled with ${user1.email}(username: ${user1.username}) right now!
+				<br>Meet Details:
+				<br>Type:${meet.type}
+				<br>Duration:${meet.duration}
+				<br>Message:${meet.msg}
+				<br>Link:${meet.link}
+				<br>Place:${meet.place}
+				`
+
+			data.mailSched(fts, {
+				'email': user1.email, 
+				'title': title, 
+				'msg': msg1
+			})
+			data.mailSched(fts, {
+				'email': user2.email, 
+				'title': title, 
+				'msg': msg2
+			})
+		}
 	} catch(err){
 		throw new Error(err);
 	}
@@ -21,7 +75,15 @@ async function meetaccept(user_id, fren_id)
 async function meetreject(user_id, fren_id)
 {
 	try{
+		errors = {}
 		var meet = await data.getMeet(user_id, fren_id);	
+		dets = await data.getUserDetsById(user_id)
+		if(dets.sendmeet.includes(meet._id))
+		{
+			errors.general = 'You have already sent the request. Kindly refresh the page.';
+			throw new UserInputError('You have already sent the request. Kindly refresh the page.', { errors });
+		}
+
 		await data.updateAccRejMeet(user_id, fren_id, meet._id, 0);
 		await data.removeNotif(user_id, fren_id, "mreq");
 		await Meet.deleteOne({_id: meet._id})
@@ -44,11 +106,17 @@ async function meetrequest(meetdata)
 		if(now > fts)
 		{
 			errors.general = 'Invalid timestamp';
-			throw new UserInputError('Invalid starting date', { errors });
+			throw new UserInputError('Invalid timestamp', { errors });
 		}
 
 		user_id = meetdata.sender
 		fren_id = meetdata.sendee
+		var oldMeet = await data.getMeet(user_id, fren_id);	
+		if(Object.keys(oldMeet).length)
+		{
+			errors.general = 'Meet thread already exists with this user. Kindly refresh page.';
+			throw new UserInputError('Meet thread already exists with this user. Kindly refresh page.', { errors });
+		}
 		await data.newNotif(fren_id, user_id, "mreq");
 
 		const newMeet = new Meet({
@@ -63,12 +131,6 @@ async function meetrequest(meetdata)
 			notif: meetdata.notif
 		});
 
-		if(meetdata.notif)
-		{
-			var task = schedule.scheduleJob('mail', moment(now).format(), 'Asia/Kolkota', function() {
-				console.log("hi")
-			})
-		}
 		await newMeet.save().then(async(saved) => {
 			await data.updateRequestMeet(user_id, fren_id, saved._id);
 			});
@@ -122,12 +184,19 @@ async function meetEdit(meetdata)
 		if(now > fts)
 		{
 			errors.general = 'Invalid timestamp';
-			throw new UserInputError('Invalid starting date', { errors });
+			throw new UserInputError('Invalid timestamp', { errors });
 		}
 
 		user_id = meetdata.sender
 		fren_id = meetdata.sendee
 		var meet = await data.getMeet(user_id, fren_id);	
+
+		dets = await data.getUserDetsById(user_id)
+		if(dets.sendmeet.includes(meet._id))
+		{
+			errors.general = 'You have already sent the request. Kindly refresh the page.';
+			throw new UserInputError('You have already sent the request. Kindly refresh the page.', { errors });
+		}
 
 		await data.removeNotif(user_id, fren_id, "mreq");
 		await data.newNotif(fren_id, user_id, "mreq");
@@ -142,14 +211,6 @@ async function meetEdit(meetdata)
 			place: meetdata.place,
 			notif: meetdata.notif
 		}});
-
-		if(meetdata.notif)
-		{
-			console.log("ni")
-			var task = schedule.scheduleJob('mail', moment(now).format(), 'Asia/Kolkota', function() {
-				console.log("hi")
-			})
-		}
 
 		await data.updateAccRejMeet(user_id, fren_id, meet._id, 0);
 		await data.updateRequestMeet(user_id, fren_id, meet._id);
